+++ app/Models/Category.php
@@ -9,7 +9,7 @@ use Illuminate\Database\Eloquent\Model;
 
 class Category extends Model
 {
-    //use HasFactory;
+    use HasFactory;
 
     use TenantTrait;
